Moving redirection logic to petitionMiddleware
