+++ middleware/redirectMiddleware.js
@@ -1,7 +1,5 @@
 const config = require('config');
 
-const crRespond = require('steps/co-respondent/cr-respond/CrRespond.step');
-
 const authTokenString = '__auth-token';
 
 const redirectOnCondition = (req, res, next) => {
@@ -11,11 +9,6 @@ const redirectOnCondition = (req, res, next) => {
   }
 
   const originalPetition = req.session.originalPetition;
-
-  const idamUserIsCorespondent = originalPetition.coRespondentAnswers && originalPetition.coRespondentAnswers.contactInfo && req.idam.userDetails.email === originalPetition.coRespondentAnswers.contactInfo.emailAddress;
-  if (idamUserIsCorespondent) {
-    return res.redirect(crRespond.path);
-  }
 
   const idamUserIsRespondent = req.idam.userDetails.email === originalPetition.respEmailAddress;
   if (!idamUserIsRespondent) {
