+++ steps/co-respondent/cr-contact-details/CrContactDetails.html
@@ -13,7 +13,7 @@
         <section class="email">
             <h2 class="heading-medium">{{ content.contactMethods.email.heading }}</h2>
             <p class="text">
-                {{ content.contactMethods.email.intro | safe }} <strong>{{ coRespEmailAddress }}</strong>.
+                {{ content.contactMethods.email.intro | safe }} <strong class="bold-small">{{ coRespEmailAddress }}</strong>.
             </p>
 
             {% call formSection() %}
