+++ test/unit/middleware/redirectMiddleware.test.js
@@ -1,6 +1,5 @@
 const { expect, sinon } = require('@hmcts/one-per-page-test-suite');
 const config = require('config');
-const crRespond = require('steps/co-respondent/cr-respond/CrRespond.step');
 
 const modulePath = 'middleware/redirectMiddleware';
 
@@ -55,30 +54,6 @@ describe(modulePath, () => {
     redirectMiddleware.redirectOnCondition(req, res, next);
 
     expect(next.calledOnce).to.eql(true);
-  });
-
-
-  it('should redirect to Co-respondent respond page if user is Co-respondent', () => {
-    const email = 'some@email.address';
-    const req = {
-      cookies: { '__auth-token': 'authToken' },
-      session: {
-        originalPetition: {
-          coRespondentAnswers: {
-            contactInfo: {
-              emailAddress: email
-            }
-          }
-        }
-      },
-      idam: {
-        userDetails: { email }
-      }
-    };
-
-    redirectMiddleware.redirectOnCondition(req, res, next);
-
-    expect(res.redirect.withArgs(crRespond.path).calledOnce).to.be.true;
   });
 
   it('should direct to DN if user is not respondent', () => {
